test/ - refactor http/https testing
- use same code for http/https testing
- http-common - new common testing
- http.test.js, https.test.js - simple wrappers
- single http directory of support files
