+++ test/probes/http/client-object.js
@@ -1,3 +1,5 @@
+'use strict';
+
 exports.data = function (ctx) {
   return {
     hostname: 'localhost',
