--- test/probes/https/query-filtering.js
@@ -1,12 +1,0 @@
-exports.data = function (ctx) {
-  return {
-    url: 'https://localhost:' + ctx.data.port + '/?foo=bar'
-  }
-}
-
-exports.run = function (ctx, done) {
-  ctx.https.get(ctx.data.url, function (res) {
-    res.resume()
-    res.on('end', done.bind(null, null))
-  }).on('error', done)
-}
